Uploaded project
The game as a whole was developed over the course of the semester, and different versions of it was submitted in different assignments on Canvas. This version is the final version with all features of the game finished, and submitted in the final assignment. Created for CSCI306 at Colorado School of Mines, and developed with Max Hampton (Rhifun). See README.txt and gitlog.txt for more information: pair programming was used throughout the game's development.